Restore scene grabbables/physics support
Scene grabbables support was reverted (https://github.com/Hubs-Foundation/hubs/pull/6515) in the Hubs client due to bugs and is now only present for the Hubs Client's addons branch and its behavior graphs add-on.  This commit restores the scene grabbables and physics support (along with fixes/improvements that were added to it in the main Blender add-on) to the behavior graphs Blender add-on to allow the functionality to be used with the Hubs Client's addons branch and its behavior graphs add-on.  A corresponding commit to the main Blender add-on removes support for scene grabbables/physics in the main Blender add-on so it maintains compatibility with the current Hubs Client.