Fix potentially adding a duplicate component when migrating the grabbable component
Prevents an extra Networked Transform component from being added to the host during the migration of the grabbable component if a Networked Transform component is somehow already present on the host.

This shouldn't generally be needed, but a duplicate Networked Transform component was added during the migration of an old blend file and adding a guard to prevent duplicates doesn't hurt.