Update Blender compatibility documentation
What: Updates the documentation on which Blender versions this add-on is compatible with.

Why: The range of Blender versions that this add-on supports has expanded greatly since it was first created, but the documentation hadn't been updated to match, which could confuse users and make them think that only a very old version of Blender is supported.